pyproject.toml
[tool.poetry]
name = "nautobot-plugin-evpn"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "nautobot_plugin_evpn"}]

[tool.poetry.dependencies]
python = "^3.10"
nautobot = "^1.4.2"
ipython = "^8.5.0"

[tool.poetry.dev-dependencies]
black = "*"
invoke = "*"
mkdocs = "^1.3.1"
pydocstyle = "^6.1.1"
pylint-django = "^2.5.3"
yamllint = "^1.28.0"
pylint = "^2.15.3"
django-debug-toolbar = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
